--- aws/context_1_7.go
@@ -1,9 +1,0 @@
-// +build go1.7
-
-package aws
-
-import "context"
-
-var (
-	backgroundCtx = context.Background()
-)
